test(pkg/cmd): fix command tests and cleanup
